accounts/abi: reorganizing package and some notes and a quick correction of name.

get rid of some imports
